Convert test file gen process for kernel region tests.
