+++ test/functional/kernel/region/CMakeLists.txt
@@ -5,15 +5,40 @@
 # SPDX-License-Identifier: (BSD-3-Clause)
 ###############################################################################
 
-raja_add_test(
-  NAME test-kernel-region-seq
-  SOURCES test-kernel-region-seq.cpp)
+list(APPEND KERNEL_REGION_BACKENDS Sequential)
 
 if(RAJA_ENABLE_OPENMP)
-  raja_add_test(
-    NAME test-kernel-region-openmp
-    SOURCES test-kernel-region-openmp.cpp)
-  raja_add_test(
-    NAME test-kernel-region-sync-openmp
-    SOURCES test-kernel-region-sync-openmp.cpp)
+  list(APPEND KERNEL_REGION_BACKENDS OpenMP)
 endif()
+
+
+#
+# Generate kernel region basic tests for each enabled RAJA back-end.
+#
+foreach( REGION_BACKEND ${KERNEL_REGION_BACKENDS} )
+  configure_file( test-kernel-region.cpp.in
+                  test-kernel-region-${REGION_BACKEND}.cpp )
+  raja_add_test( NAME test-kernel-region-${REGION_BACKEND}
+                 SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-kernel-region-${REGION_BACKEND}.cpp )
+
+  target_include_directories(test-kernel-region-${REGION_BACKEND}.exe
+                             PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
+endforeach()
+
+unset( KERNEL_REGION_BACKENDS )
+
+
+#
+# Generate kernel region sync tests for OpenMP.
+#
+set(REGION_BACKEND OpenMP)
+
+configure_file( test-kernel-region-sync.cpp.in
+                test-kernel-region-sync-${REGION_BACKEND}.cpp )
+raja_add_test( NAME test-kernel-region-sync-${REGION_BACKEND}
+               SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-kernel-region-sync-${REGION_BACKEND}.cpp )
+
+target_include_directories(test-kernel-region-sync-${REGION_BACKEND}.exe
+                           PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
+
+unset(REGION_BACKEND)
