--- test/functional/kernel/region/test-kernel-region-seq.cpp
@@ -1,52 +1,0 @@
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
-// Copyright (c) 2016-20, Lawrence Livermore National Security, LLC
-// and RAJA project contributors. See the RAJA/COPYRIGHT file for details.
-//
-// SPDX-License-Identifier: (BSD-3-Clause)
-//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
-
-#include "test-kernel-region.hpp"
-
-using SequentialKernelRegionExecPols = 
-  camp::list< 
-
-    RAJA::KernelPolicy<
-      RAJA::statement::Region<RAJA::seq_region,
-        RAJA::statement::For<0, RAJA::seq_exec,
-          RAJA::statement::Lambda<0>
-        >,
-        RAJA::statement::For<0, RAJA::seq_exec,
-          RAJA::statement::Lambda<1>
-        >,
-        RAJA::statement::For<0, RAJA::seq_exec,
-          RAJA::statement::Lambda<2>
-        >
-      >
-    >,
-
-    RAJA::KernelPolicy<
-      RAJA::statement::Region<RAJA::seq_region,
-        RAJA::statement::For<0, RAJA::loop_exec,
-          RAJA::statement::Lambda<0>
-        >,
-        RAJA::statement::For<0, RAJA::loop_exec,
-          RAJA::statement::Lambda<1>
-        >,
-        RAJA::statement::For<0, RAJA::loop_exec,
-          RAJA::statement::Lambda<2>
-        >
-      >
-    >
-
-  >;
-
-
-// Cartesian product of types for Sequential tests
-using SequentialKernelRegionTypes =
-  Test< camp::cartesian_product<IdxTypeList, 
-                                HostResourceList,
-                                SequentialKernelRegionExecPols> >::Types;
-
-INSTANTIATE_TYPED_TEST_SUITE_P(Sequential,
-                               KernelRegionBasicTest,
-                               SequentialKernelRegionTypes);
